test: split keymap-dispatch edge suites out of the oversized test file
keymap-dispatch.test.ts crossed the file-size cap at 749 lines once the
ctrl-hold coverage landed. The modal-barrier, shadowed-match-warning, and
shift+letter describes now live in keymap-dispatch-edges.test.ts with the
shared makeEvt/makeReg helpers; both files sit under the cap.
